fix merge compile errors
