+++ bevy_mod_scripting_core/Cargo.toml
@@ -32,6 +32,6 @@ bevy = { version = "0.8.0", default-features = false, features=["bevy_asset","be
 bevy_event_priority = {path = "../bevy_event_priority", version = "0.1.1"}
 thiserror = "1.0.31"
 paste = "1.0.7"
-
+parking_lot = "0.12.1"
