+++ bevy_script_api/src/lib.rs
@@ -3,6 +3,9 @@ extern crate bevy;
 pub mod error;
 #[cfg(feature = "lua")]
 pub mod lua;
+#[cfg(feature= "rhai")]
+pub mod rhai;
+
 pub mod script_ref;
 pub mod sub_reflect;
 pub mod wrappers;
