+++ bevy_script_api/src/rhai/bevy/mod.rs
@@ -1,7 +1,8 @@
 use std::ops::{Deref, DerefMut};
-use rhai::plugin::*;
 
-use crate::{common::bevy::WorldPointer, APIProvider, RhaiContext, RhaiDocFragment};
+use bevy_mod_scripting_core::{prelude::*, world::WorldPointer};
+use bevy_mod_scripting_rhai::{rhai as rhai, prelude::*};
+use rhai::plugin::*;
 
 #[derive(Clone)]
 pub struct RhaiWorld(WorldPointer);
@@ -33,11 +34,9 @@ impl RhaiWorld {
     }
 }
 
-#[::rhai::export_module]
+#[rhai::export_module]
 pub mod bevy_plugin{
-    use crate::api::rhai::bevy::RhaiWorld;
     pub type World = RhaiWorld;
-
 }
 
 
@@ -48,7 +47,7 @@ impl APIProvider for RhaiBevyAPIProvider {
     type ScriptContext=RhaiContext;
     type DocTarget=RhaiDocFragment;
 
-    fn attach_api(&mut self, engine: &mut Self::APITarget) -> Result<(), crate::ScriptError> {
+    fn attach_api(&mut self, engine: &mut Self::APITarget) -> Result<(), ScriptError> {
         engine.register_global_module(exported_module!(bevy_plugin).into());
         Ok(())
     }
