+++ bevy_script_api/src/sub_reflect.rs
@@ -1,13 +1,18 @@
 use std::{borrow::Cow, sync::Weak};
+use std::fmt;
+use std::fmt::{Debug, Display};
+use parking_lot::RwLock;
 
-use crate::error::ReflectionError;
 use bevy::{
     prelude::{Entity, ReflectComponent, ReflectResource, World},
     reflect::{Reflect, ReflectMut, ReflectRef},
 };
-use parking_lot::RwLock;
 
-use std::fmt;
+use bevy_mod_scripting_core::error::ScriptError;
+use crate::{script_ref::ReflectPtr};
+use crate::error::ReflectionError;
+use bevy_mod_scripting_core::world::WorldPointer;
+
 
 /// The base of a reflect path, i.e. the top-level object or source. Reflections paths are always relative to some reflect base
 #[derive(Clone)]
@@ -102,10 +107,7 @@ pub enum ReflectPathElem {
     /// Access to a TupleStruct, Tuple, List or Array element
     IndexAccess(usize), // TODO: Map access
 }
-use std::fmt::{Debug, Display};
 
-use bevy_mod_scripting_core::error::ScriptError;
-use crate::{script_ref::ReflectPtr, common::bevy::WorldPointer};
 
 impl Debug for ReflectPathElem {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
